[Backport 2023.2] [HDRP] Bunch of fixes for HDRP (Version 20)
gathering of PRs

Fixed: [HDRP] realtime Reflection probe makes volumetrics clouds wind stop (#30226)
Avoid Allocating Decal Atlas When Disabled (#31528) 
Fix water editor no pipeline (#31761)
Fix refraction preview (#31797)
Use floats for min and max nits (#31990)
Only enable hdr at startup (#31988)
Graphics/hd/uum40657 (#31967) 
Fix GraphicsBuffer Leak in APV Binding Code (#31887)
Re-Enable HDR output on Mac (#33640)

+++ Packages/com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipeline.PostProcess.cs
@@ -4462,8 +4462,8 @@ namespace UnityEngine.Rendering.HighDefinition
         static void GetHDROutputParameters(HDROutputUtils.HDRDisplayInformation hdrDisplayInformation, ColorGamut hdrDisplayColorGamut, Tonemapping tonemappingComponent, out Vector4 hdrOutputParameters1, out Vector4 hdrOutputParameters2)
         {
             ColorGamut gamut = hdrDisplayColorGamut;
-            var minNits = hdrDisplayInformation.minToneMapLuminance;
-            var maxNits = hdrDisplayInformation.maxToneMapLuminance;
+            var minNits = (float)hdrDisplayInformation.minToneMapLuminance;
+            var maxNits = (float)hdrDisplayInformation.maxToneMapLuminance;
             var paperWhite = hdrDisplayInformation.paperWhiteNits;
             int eetfMode = 0;
             float hueShift = 0.0f;
@@ -4509,8 +4509,8 @@ namespace UnityEngine.Rendering.HighDefinition
             }
             if (!tonemappingComponent.detectBrightnessLimits.value)
             {
-                minNits = (int)tonemappingComponent.minNits.value;
-                maxNits = (int)tonemappingComponent.maxNits.value;
+                minNits = tonemappingComponent.minNits.value;
+                maxNits = tonemappingComponent.maxNits.value;
             }
 
             hdrOutputParameters1 = new Vector4(minNits, maxNits, paperWhite, 1f / paperWhite);
